created data folder and moved TeachersData into it
